% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_graph_list}
\alias{generate_graph_list}
\title{Generates a bigger graph using either linear or exponential growth.}
\usage{
generate_graph_list(
  num_graphs = 15,
  del_edge = 0.1,
  new_nodes = 0.1,
  edge_increase = 0.1,
  mode = "exp"
)
}
\arguments{
\item{num_graphs}{The number of graphs to be generated.
Default set to \code{15}.}

\item{del_edge}{The proportion of edges deleted from the input graph. Default
set to \code{0.1}.}

\item{new_nodes}{The proportion of nodes added to the input graph. Default
set to \code{0.1}.}

\item{edge_increase}{The proportion of edges added to the input graph. Default
set to \code{0.1}.}

\item{mode}{The method the graphs grow in.
Either \code{"linear"} for linear creation, or \code{"exp"}.
Default set to \code{"exp"}.}

\item{gr}{The input graph to generate the next graph. If set to \code{NULL}
a graph using \code{igraph::sample_pa} is used as the input graph.}
}
\value{
A graph list.
}
\description{
Wrapper around 'generate_graph_linear' and 'generate_graph_exp'.
}
\examples{
set.seed(1)
gr <- generate_graph_list(12, 0.1, 0.1, 0.1, "exp")
gr

}
