% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syngraphs-data.R
\docType{data}
\name{syngraphs}
\alias{syngraphs}
\title{A dataset containing synthetic graphs}
\format{
A list of 20 igraph graphs
}
\usage{
syngraphs
}
\description{
This dataset contains a list of synthetic igraph objects
}
\examples{
data(syngraphs)
syngraphs[[1]]
}
\keyword{datasets}
